+++ demo/src/app/pages/message-dialog-consumer.component.ts
@@ -15,16 +15,17 @@
  */
 
 import { Component } from '@angular/core';
-import { ErrorDialogExample1Component } from '../code-snippets/error-dialog/example-1.component';
+import { MessageDialogExample1Component } from '../code-snippets/message-dialog-service/example-1.component';
 import { CodeExampleBoxComponent, CodeExampleBoxComponentOptions } from '../utils/code-example-box.component';
 import { CodeSnippetBoxComponent } from '../utils/code-snippet-box.component';
 import { DocTableComponent, DocTableRowData } from '../utils/doc-table.component';
 import { ExpandableSectionsComponent } from '../utils/expandable-sections.component';
 import { OverviewSectionComponent } from '../utils/overview-section.component';
 import { SubsectionTitleComponent } from '../utils/subsection-title.component';
+import { MessageDialogExample2Component } from '../code-snippets/message-dialog-service/example-2.component';
 
 @Component({
-  selector: 'app-error-dialog',
+  selector: 'app-message-dialog',
   imports: [
     OverviewSectionComponent,
     SubsectionTitleComponent,
@@ -32,7 +33,9 @@ import { SubsectionTitleComponent } from '../utils/subsection-title.component';
     ExpandableSectionsComponent,
     CodeExampleBoxComponent,
     CodeSnippetBoxComponent,
-    ErrorDialogExample1Component,
+    MessageDialogExample1Component,
+    MessageDialogExample2Component,
+    MessageDialogExample2Component,
   ],
   template: `
     <app-expandable-sections>
@@ -53,23 +56,32 @@ import { SubsectionTitleComponent } from '../utils/subsection-title.component';
         <app-subsection-title>showErrorDialog - Parameters</app-subsection-title>
         <app-doc-table [dataSource]="showErrorDialogDocTableDataSource"></app-doc-table>
 
+        <app-subsection-title>showSubmitDialog - Parameters</app-subsection-title>
+        <app-doc-table [dataSource]="showSubmitDialogDocTableDataSource"></app-doc-table>
+
         <app-subsection-title>Interface ErrorsDialogProps - Properties</app-subsection-title>
         <app-doc-table [dataSource]="errorsDialogPropsDocTableDataSource"></app-doc-table>
+
+        <app-subsection-title>Interface SubmitDialogProps - Properties</app-subsection-title>
+        <app-doc-table [dataSource]="submitDialogPropsDocTableDataSource"></app-doc-table>
 
         <app-subsection-title>Interface ErrorMessage - Properties</app-subsection-title>
         <app-doc-table [dataSource]="errorMessageDocTableDataSource"></app-doc-table>
 
-        <app-subsection-title>Interface ErrorDialogStyle - Properties</app-subsection-title>
-        <app-doc-table [dataSource]="errorDialogStyleDocTableDataSource"></app-doc-table>
+        <app-subsection-title>Interface DialogStyle - Properties</app-subsection-title>
+        <app-doc-table [dataSource]="dialogStyleDocTableDataSource"></app-doc-table>
       </app-overview-section>
 
       <app-code-example-box [options]="examples[0]">
-        <app-error-dialog-example-1></app-error-dialog-example-1>
+        <app-message-dialog-example-1></app-message-dialog-example-1>
+      </app-code-example-box>
+      <app-code-example-box [options]="examples[1]">
+        <app-message-dialog-example-2></app-message-dialog-example-2>
       </app-code-example-box>
     </app-expandable-sections>
   `,
 })
-export class ErrorDialogConsumerComponent {
+export class MessageDialogConsumerComponent {
   methodsDocTableDataSource: DocTableRowData[] = [
     {
       name: '`showErrorDialog`',
@@ -78,28 +90,68 @@ export class ErrorDialogConsumerComponent {
         'Each error can have a search query attached. If so, the error message contains a link button to direct the user to the DEMIS Knowledge Base.',
       ],
     },
+    {
+      name: '`showSubmitDialog`',
+      description: [
+        'Opens a successful confirmation dialog after submitting a notification.',
+        'Shows notification details including ID, timestamp, and provides a PDF download link.',
+        'The dialog automatically triggers a PDF download and provides navigation back to the homepage.',
+        'The dialog cannot be closed by clicking outside or pressing ESC (disableClose: true).',
+      ],
+    },
   ];
 
   showErrorDialogDocTableDataSource: DocTableRowData[] = [
     { name: '`data: ErrorsDialogProps`', description: 'The data object used to render the error dialog.' },
-    { name: '`style?: ErrorDialogStyle`', description: '`[optional]` Overwrite styles of the message dialog.' },
+    { name: '`style?: DialogStyle`', description: '`[optional]` Overwrite styles of the message dialog.' },
+  ];
+
+  showSubmitDialogDocTableDataSource: DocTableRowData[] = [
+    { name: '`data: SubmitDialogProps`', description: 'The data object used to render the submit success dialog.' },
+    {
+      name: '`style?: DialogStyle`',
+      description: '`[optional]` Overwrite styles of the message dialog. Note: disableClose is always set to true.',
+    },
   ];
 
   errorsDialogPropsDocTableDataSource: DocTableRowData[] = [
-    { name: '`errors: ErrorMessage[]`', description: 'The array of error messages to be displayed in the error dialog.' },
+    {
+      name: '`errors: ErrorMessage[]`',
+      description: 'The array of error messages to be displayed in the error dialog.',
+    },
     { name: '`clipboardContent?: string`', description: '`[optional]`' },
     { name: '`errorTitle?: string`', description: '`[optional]` Overwrites the default dialog title.' },
   ];
 
-  errorMessageDocTableDataSource: DocTableRowData[] = [
-    { name: '`text: string`', description: 'The error text message to be displayed in the error dialog.' },
-    { name: '`queryString?: string`', description: '`[optional]` A search string that will trigger a button to direct the user to the DEMIS Knowledge Base.' },
+  submitDialogPropsDocTableDataSource: DocTableRowData[] = [
+    { name: '`notificationId: string`', description: 'The unique identifier for the submitted notification.' },
+    { name: '`timestamp: string`', description: 'The formatted timestamp when the notification was submitted.' },
+    { name: '`fileName: string`', description: 'The name of the PDF file to be downloaded.' },
+    { name: '`href: string`', description: 'The download URL for the PDF confirmation file.' },
+    { name: '`authorEmail: string`', description: 'The email address for support contact.' },
   ];
 
-  errorDialogStyleDocTableDataSource: DocTableRowData[] = [
-    { name: '`height?: string`', description: '`[optional]` Overwrites the height config values of the Material Dialog used.' },
-    { name: '`width?: string`', description: '`[optional]` Overwrites the width config values of the Material Dialog used.' },
-    { name: '`maxWidth?: string`', description: '`[optional]` Overwrites the maxWidth config values of the Material Dialog used.' },
+  errorMessageDocTableDataSource: DocTableRowData[] = [
+    { name: '`text: string`', description: 'The error text message to be displayed in the error dialog.' },
+    {
+      name: '`queryString?: string`',
+      description: '`[optional]` A search string that will trigger a button to direct the user to the DEMIS Knowledge Base.',
+    },
+  ];
+
+  dialogStyleDocTableDataSource: DocTableRowData[] = [
+    {
+      name: '`height?: string`',
+      description: '`[optional]` Overwrites the height config values of the Material Dialog used.',
+    },
+    {
+      name: '`width?: string`',
+      description: '`[optional]` Overwrites the width config values of the Material Dialog used.',
+    },
+    {
+      name: '`maxWidth?: string`',
+      description: '`[optional]` Overwrites the maxWidth config values of the Material Dialog used.',
+    },
   ];
 
   examples: CodeExampleBoxComponentOptions[] = [
@@ -110,12 +162,28 @@ export class ErrorDialogConsumerComponent {
         {
           fileName: 'example-1.component.html',
           language: 'html',
-          codeSnippetPath: 'code-snippets/error-dialog',
+          codeSnippetPath: 'code-snippets/message-dialog-service',
         },
         {
           fileName: 'example-1.component.ts',
           language: 'ts',
-          codeSnippetPath: 'code-snippets/error-dialog',
+          codeSnippetPath: 'code-snippets/message-dialog-service',
+        },
+      ],
+    },
+    {
+      expanderTitle: 'Example 2',
+      expanderDescription: 'A simple submit dialog',
+      codeSnippets: [
+        {
+          fileName: 'example-2.component.html',
+          language: 'html',
+          codeSnippetPath: 'code-snippets/message-dialog-service',
+        },
+        {
+          fileName: 'example-2.component.ts',
+          language: 'ts',
+          codeSnippetPath: 'code-snippets/message-dialog-service',
         },
       ],
     },